requirements.txt
alpaca-py>=0.21.0
pandas>=2.2.0
python-dotenv>=1.0.1
pydantic>=2.7.0
matplotlib>=3.8.0
